+++ src/main/java/com/softserve/itacademy/kek/repositories/IdentityTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IdentityTypeRepository extends CrudRepository<IdentityType, Integer> {
+public interface IdentityTypeRepository extends CrudRepository<IdentityType, Long> {
 }
